Close #51. Add $0 goals when no value has been assigned.
Another way to do it would be to allow the report_goal_lines to show up
in the reports and in the lists even when the goals aren't there, but
that would take more work.

I'm going to commit this to a separate branch and wait for feedback
before submitting this.